[EpoxyBars] Don't apply insets when hierarchy has a 3d transform
If any view in the hierarchy has a 3D transform, it's not valid to apply the insets as they may be incorrect; we should wait until there is no transform to so do.
