+++ Sources/EpoxyBars/BottomBarInstaller/BottomBarContainer.swift
@@ -160,6 +160,10 @@ public final class BottomBarContainer: BarStackView, InternalBarContainer {
   /// Updates the view controller insets (either safe area or scroll view content inset) in response
   /// to the safe area, center, or bounds changing.
   private func updateInsets() {
+    // If any view in the hierarchy has a 3D transform, it's not valid to apply the insets as they
+    // may be incorrect; we should wait until there is no transform to so do.
+    guard !hasHierarchy3DTransform() else { return }
+
     updateAdditionalSafeAreaInset(additionalSafeAreaInsetsBottom)
 
     // If offset from the bottom, use the original layout margins rather than the safe area margins,
@@ -169,13 +173,4 @@ public final class BottomBarContainer: BarStackView, InternalBarContainer {
     setLayoutMargin(margin)
   }
 
-}
-
-// MARK: - UIViewController
-
-extension UIViewController {
-  @nonobjc
-  fileprivate var originalSafeAreaInsetBottom: CGFloat {
-    view.safeAreaInsets.bottom - additionalSafeAreaInsets.bottom
-  }
 }
